Change redux architecture to fetch on demand Change redux state architecture to fetch/resolve items on demand
Our data architecture was modelled around a single view that holds items found nearby. With the addtional of deep linking and the ability to render items that may not be nearby we needed to decouple 'nearby items' from 'items' in our app state and provide way to load item data when an item hasn't come from the scanner.